feat: dynamic configuration for caddy

+++ app/Livewire/Server/Proxy/DynamicConfigurationNavbar.php
@@ -15,8 +15,16 @@ class DynamicConfigurationNavbar extends Component
     {
         $server = Server::ownedByCurrentTeam()->whereId($this->server_id)->first();
         $proxy_path = $server->proxyPath();
+        $proxy_type = $server->proxyType();
         $file = str_replace('|', '.', $fileName);
+        if ($proxy_type === 'CADDY' && $file === "Caddyfile") {
+            $this->dispatch('error', 'Cannot delete Caddyfile.');
+            return;
+        }
         instant_remote_process(["rm -f {$proxy_path}/dynamic/{$file}"], $server);
+        if ($proxy_type === 'CADDY') {
+            $server->reloadCaddy();
+        }
         $this->dispatch('success', 'File deleted.');
         $this->dispatch('loadDynamicConfigurations');
         $this->dispatch('refresh');
